+++ randovania/game_description/pickup/pickup_migration.py
@@ -4,6 +4,7 @@ import itertools
 
 from randovania.game.game_enum import RandovaniaGame
 from randovania.game_description import migration_data
+from randovania.layout.base.standard_pickup_state import StartingPickupBehavior
 from randovania.lib import migration_lib
 
 
@@ -199,6 +200,21 @@ def _migrate_v12(pickup_data: dict, game: RandovaniaGame) -> None:
         category.pop("is_broad_category", None)
 
 
+def _migrate_v13(pickup_data: dict, game: RandovaniaGame) -> None:
+    standard_pickups = pickup_data["standard_pickups"]
+    for pickup in standard_pickups.values():
+        if "must_be_starting" in pickup:
+            pickup["starting_condition"] = StartingPickupBehavior.MUST_BE_STARTING
+            pickup.pop("must_be_starting")
+        else:
+            pickup["starting_condition"] = StartingPickupBehavior.CAN_BE_STARTING
+    if game == RandovaniaGame.METROID_PRIME_ECHOES:
+        banned_starting_items = ["Cannon Ball", "Double Damage", "Unlimited Beam Ammo", "Unlimited Missiles"]
+        for pickup in standard_pickups.keys():
+            if pickup in banned_starting_items:
+                standard_pickups[pickup]["starting_condition"] = StartingPickupBehavior.CAN_NEVER_BE_STARTING
+
+
 _MIGRATIONS = [
     None,
     None,
@@ -212,6 +228,7 @@ _MIGRATIONS = [
     _migrate_v10,  # move category fields to pickup and add hint_features
     _migrate_v11,  # fix the fact that old migrations don't actually work on old DBs, lol
     _migrate_v12,  # remove is_broad_category
+    _migrate_v13,  # change must_be_starting to starting_condition
 ]
 CURRENT_VERSION = migration_lib.get_version(_MIGRATIONS)