+++ randovania/games/blank/gui/preset_settings/__init__.py
@@ -14,6 +14,7 @@ def preset_tabs(editor: PresetEditor, window_manager: WindowManager) -> list[typ
     from randovania.games.blank.gui.preset_settings.blank_patches_tab import PresetBlankPatches
     from randovania.gui.preset_settings.dock_rando_tab import PresetDockRando
     from randovania.gui.preset_settings.generation_tab import PresetGeneration
+    from randovania.gui.preset_settings.hints_tab import PresetHints
     from randovania.gui.preset_settings.item_pool_tab import PresetItemPool
     from randovania.gui.preset_settings.location_pool_tab import PresetLocationPool
     from randovania.gui.preset_settings.trick_level_tab import PresetTrickLevel
@@ -21,6 +22,7 @@ def preset_tabs(editor: PresetEditor, window_manager: WindowManager) -> list[typ
     return [
         PresetTrickLevel,
         PresetGeneration,
+        PresetHints,
         PresetLocationPool,
         PresetItemPool,
         PresetBlankPatches,