+++ randovania/games/fusion/game_data.py
@@ -6,6 +6,7 @@ import randovania.game.data
 import randovania.game.development_state
 import randovania.game.generator
 import randovania.game.gui
+import randovania.game.hints
 import randovania.game.layout
 from randovania.games.fusion import layout
 from randovania.games.fusion.db_integrity import find_fusion_db_errors
@@ -39,15 +40,31 @@ def _gui() -> randovania.game.gui.GameGui:
 
 def _generator() -> randovania.game.generator.GameGenerator:
     from randovania.games.fusion import generator
-    from randovania.games.fusion.generator.hint_distributor import FusionHintDistributor
     from randovania.generator.filler.weights import ActionWeights
 
     return randovania.game.generator.GameGenerator(
         pickup_pool_creator=generator.pool_creator,
         bootstrap=generator.FusionBootstrap(),
         base_patches_factory=generator.FusionBasePatchesFactory(),
-        hint_distributor=FusionHintDistributor(),
         action_weights=ActionWeights(events_weight=0.75, hints_weight=0.5),
+    )
+
+
+def _hints() -> randovania.game.hints.GameHints:
+    from randovania.games.fusion.generator.hint_distributor import FusionHintDistributor
+
+    return randovania.game.hints.GameHints(
+        hint_distributor=FusionHintDistributor(),
+        specific_pickup_hints={
+            "artifacts": randovania.game.hints.SpecificHintDetails(
+                long_name="Infant Metroids",
+                description="This controls how precise the Infant Metroids hint at Restricted Labs is.",
+            ),
+            "charge_beam": randovania.game.hints.SpecificHintDetails(
+                long_name="Charge Beam",
+                description="This controls how precise the hint for Charge Beam in Auxiliary Navigation Room is.",
+            ),
+        },
     )
 
 
@@ -126,6 +143,7 @@ game_data: randovania.game.data.GameData = randovania.game.data.GameData(
     options=_options,
     gui=_gui,
     generator=_generator,
+    hints=_hints,
     patch_data_factory=_patch_data_factory,
     exporter=_exporter,
     multiple_start_nodes_per_area=True,