+++ randovania/games/fusion/layout/fusion_configuration.py
@@ -5,7 +5,6 @@ import dataclasses
 from randovania.bitpacking.bitpacking import BitPackDataclass
 from randovania.bitpacking.json_dataclass import JsonDataclass
 from randovania.game.game_enum import RandovaniaGame
-from randovania.games.fusion.layout.hint_configuration import HintConfiguration
 from randovania.layout.base.base_configuration import BaseConfiguration
 
 
@@ -23,7 +22,6 @@ class FusionConfiguration(BaseConfiguration):
     instant_transitions: bool
     energy_per_tank: int = dataclasses.field(metadata={"min": 1, "max": 1000, "precision": 1})
     artifacts: FusionArtifactConfig
-    hints: HintConfiguration
     open_save_recharge_hatches: bool
     short_intro_text: bool