+++ randovania/games/prime2/game_data.py
@@ -4,6 +4,7 @@ import randovania.game.data
 import randovania.game.development_state
 import randovania.game.generator
 import randovania.game.gui
+import randovania.game.hints
 import randovania.game.layout
 import randovania.game.web_info
 from randovania.games.prime2.layout.echoes_configuration import EchoesConfiguration
@@ -43,7 +44,6 @@ def _gui() -> randovania.game.gui.GameGui:
 def _generator() -> randovania.game.generator.GameGenerator:
     from randovania.games.prime2.generator.base_patches_factory import EchoesBasePatchesFactory
     from randovania.games.prime2.generator.bootstrap import EchoesBootstrap
-    from randovania.games.prime2.generator.hint_distributor import EchoesHintDistributor
     from randovania.games.prime2.generator.pickup_pool.pool_creator import echoes_specific_pool
     from randovania.generator.filler.weights import ActionWeights
 
@@ -51,8 +51,21 @@ def _generator() -> randovania.game.generator.GameGenerator:
         pickup_pool_creator=echoes_specific_pool,
         bootstrap=EchoesBootstrap(),
         base_patches_factory=EchoesBasePatchesFactory(),
-        hint_distributor=EchoesHintDistributor(),
         action_weights=ActionWeights(),
+    )
+
+
+def _hints() -> randovania.game.hints.GameHints:
+    from randovania.games.prime2.generator.hint_distributor import EchoesHintDistributor
+
+    return randovania.game.hints.GameHints(
+        hint_distributor=EchoesHintDistributor(),
+        specific_pickup_hints={
+            "sky_temple_keys": randovania.game.hints.SpecificHintDetails(
+                long_name="Sky Temple Keys",
+                description="This controls how precise the hints for Sky Temple Keys in Sky Temple Gateway are.",
+            )
+        },
     )
 
 
@@ -165,6 +178,7 @@ Taking the transport hologram at the center of this room.""",
     options=_options,
     gui=_gui,
     generator=_generator,
+    hints=_hints,
     patch_data_factory=_patch_data_factory,
     exporter=_exporter,
     defaults_available_in_game_sessions=True,